Time: 29 ms (5.04%), Space: 61.9 MB (87.05%) - LeetHub

+++ 3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.java
@@ -1,8 +1,8 @@
 class Solution {
     public long maxTotalValue(int[] nums, int k) {
         Arrays.sort(nums);
-        long a=nums[0];
-        long b=nums[nums.length-1];
-        return (b-a)*k; 
+        int a=nums[0];
+        int b=nums[nums.length-1];
+        return (long)(b-a)*k; 
     }
 }
